Extract ListeningThread to its own file
and SlimBufferingResponseListener to its own class
